requirements.txt
numpy<1.19.0
gym
keras
tensorflow==2.3.1
matplotlib
opencv-python
pillow==7.2.0
atari-py
h5py<3.0.0
python-dotenv